feat: Add map link to change events table
Adds a new column to the change event table that displays a link to Google Maps for events with location data. This allows users to quickly view the geographical location of a specific event. The row click handler has been renamed to reflect its new purpose of opening a map link.